Handle constant classOf types.
Scalac encoded `classOf[T]` singleton types as `ConstantType(Type)`.
This commit adds custom handling for this type in metacp and metac
encoding them instead their widened representation `Class[T]`.

Fixes #1306
