+++ semanticdb/integration/src/main/scala/example/Types.scala
@@ -104,5 +104,6 @@ object Test {
     final val bool = true
     final val unit = ()
     final val javaEnum = java.nio.file.LinkOption.NOFOLLOW_LINKS
+    final val clazzOf = classOf[Option[Int]]
   }
 }
